Bring selection logic to Mus7afAssignmentScreen
