+++ Dev/src/screens/MushafScreen/MushafScreen.js
@@ -6,15 +6,11 @@ import colors from "config/colors";
 import QcParentScreen from 'screens/QcParentScreen';
 import SelectionPage from "./Components/SelectionPage";
 import Swiper from "react-native-swiper";
-import { compareOrder } from "./Helpers/AyahsOrder";
-import surahs from "./Data/Surahs.json";
 import strings from "config/strings";
-import fontStyles from "config/fontStyles";
 import FirebaseFunctions from "config/FirebaseFunctions";
 import studentImages from "config/studentImages";
 import classImages from 'config/classImages';
 import { screenHeight, screenWidth } from "config/dimensions";
-import SwitchSelector from 'react-native-switch-selector';
 
 //------- constants to indicate the case when there is no ayah selected
 const noAyahSelected = {
@@ -60,9 +56,7 @@ export default class MushafScreen extends QcParentScreen {
         ayah: 0,
       },
     },
-    assignmentName: this.props.assignmentName,
     assignmentType: this.props.assignmentType,
-    freeFormAssignment: false,
     loadScreenOnClose: this.props.loadScreenOnClose,
     popOnClose: this.props.popOnClose,
     isLoading: true,
@@ -80,7 +74,7 @@ export default class MushafScreen extends QcParentScreen {
     } = this.props;
 
     //we mimmic right to left pages scanning by reversing the pages order in the swiper component
-    allPages = Array.from(Array(604), (e,i)=>604-i);
+    allPages = Array.from(Array(604), (e, i) => 604 - i);
 
     if (studentID === undefined) {
       //assign to all class -----
@@ -102,7 +96,6 @@ export default class MushafScreen extends QcParentScreen {
           imageID: currentClass.classImageID,
           classID,
           currentClass: currentClass,
-          assignmentName: currentClass.currentAssignment,
           assignmentType: currentClass.currentAssignmentType,
         },
         () => {
@@ -137,7 +130,6 @@ export default class MushafScreen extends QcParentScreen {
               completed: true,
             }
           : noSelection,
-        assignmentName,
         assignmentType:
           assignmentType !== undefined ? assignmentType : strings.Memorization
       });
@@ -196,142 +188,6 @@ export default class MushafScreen extends QcParentScreen {
   }
 
   // ------------------------- Event handlers --------------------------------------------
-
-  //this is to update the assignment text without mapping it to a location in the mus7af
-  // this is to allow teachers to enter free form assignemnts
-  // for example: redo your last 3 assignments
-  setFreeFormAssignmentName(freeFormAssignmentName) {
-    this.setState({
-      assignmentName: freeFormAssignmentName,
-      freeFormAssignment: true,
-    });
-  }
-
-  onSelectAyah(selectedAyah) {
-    const { selection } = this.state;
-
-    //if the user taps on the same selected aya again, turn off selection
-    if (
-      compareOrder(selection.start, selection.end) === 0 &&
-      compareOrder(selection.start, selectedAyah) === 0
-    ) {
-      this.setState({ selection: noSelection }, () =>
-        this.updateAssignmentName()
-      );
-    } else if (!selection.started) {
-      this.setState(
-        {
-          selection: {
-            started: true,
-            completed: false,
-            start: selectedAyah,
-            end: selectedAyah,
-          }
-        },
-        () => this.updateAssignmentName()
-      );
-    } else if (!selection.completed) {
-      this.setState(
-        {
-          selection: {
-            ...this.state.selection,
-            started: false,
-            completed: true,
-          },
-        },
-        () => {
-          //Set the smallest number as the start, and the larger as the end
-          if (compareOrder(selection.start, selectedAyah) > 0) {
-            this.setState(
-              { selection: { ...this.state.selection, end: selectedAyah } },
-              () => this.updateAssignmentName()
-            );
-          } else {
-            this.setState(
-              { selection: { ...this.state.selection, start: selectedAyah } },
-              () => this.updateAssignmentName()
-            );
-          }
-        }
-      );
-    }
-  }
-
-  updateAssignmentName() {
-    const { selection } = this.state;
-    if (selection.start.surah === 0) {
-      //no selection made
-      //todo: make this an explicit flag
-      return '';
-    }
-
-    desc = surahs[selection.start.surah].tname + " (" + selection.start.ayah;
-
-    if (selection.start.surah === selection.end.surah) {
-      if (selection.start.ayah !== selection.end.ayah) {
-        desc += strings.To + selection.end.ayah;
-      }
-    } else {
-      desc +=
-        ")" +
-        strings.To +
-        surahs[selection.end.surah].tname +
-        " (" +
-        selection.end.ayah;
-    }
-
-    let pageDesc = strings.ParenthesisPage + selection.end.page;
-    if (selection.start.page !== selection.end.page) {
-      pageDesc =
-        strings.PagesWithParenthesis +
-        selection.start.page +
-        strings.To +
-        selection.end.page;
-    }
-    desc += pageDesc;
-
-    this.setState({
-      assignmentName: desc,
-      freeFormAssignment: false,
-    });
-  }
-
-  // ---- selects a range of ayahs   -----
-  onSelectAyahs(firstAyah, lastAyah) {
-    let startA = firstAyah;
-    let endA = lastAyah;
-
-    //Set the smallest number as the start, and the larger as the end
-    if (compareOrder(firstAyah, lastAyah) <= 0) {
-      startA = lastAyah;
-      endA = firstAyah;
-    }
-
-    this.setState(
-      prevState => {
-        return {
-          selection: {
-            ...prevState.selection,
-            start: startA
-          }
-        };
-      },
-      () =>
-        this.setState(
-          prevState2 => {
-            return {
-              selection: {
-                ...prevState2.selection,
-                started: false,
-                completed: true,
-                end: endA,
-              },
-            };
-          },
-          () => this.updateAssignmentName()
-        )
-    );
-  }
 
   onChangePage(page, keepSelection) {
     //reset the selection state if we are passed a flag to do so
@@ -413,13 +269,15 @@ export default class MushafScreen extends QcParentScreen {
           //callback when user taps on a single ayah to selects
           //determines whether this would be the start of end of the selection
           // and select ayahs in between
-          onSelectAyah={this.onSelectAyah.bind(this)}
+          onSelectAyah={selectedAyah => this.props.onSelectAyah(selectedAyah)}
           //callback when user selects a range of ayahs (line an entire page or surah)
-          onSelectAyahs={this.onSelectAyahs.bind(this)}
+          onSelectAyahs={(firstAyah, lastAyah) =>
+            this.props.onSelectAyahs(firstAyah, lastAyah)
+          }
           topRightIconName={this.props.topRightIconName}
           topRightOnPress={this.props.topRightOnPress}
           onUpdateAssignmentName={newAssignmentName =>
-            this.setFreeFormAssignmentName(newAssignmentName)
+            this.props.setFreeFormAssignmentName(newAssignmentName)
           }
         />
       </View>
@@ -427,7 +285,7 @@ export default class MushafScreen extends QcParentScreen {
   }
 
   render() {
-    const { isLoading, assignmentName, currentClass } = this.state;
+    const { isLoading } = this.state;
 
     if (isLoading === true) {
       return (
@@ -439,24 +297,6 @@ export default class MushafScreen extends QcParentScreen {
         </View>
       );
     } else {
-      const options = [
-        { label: strings.Memorization, value: strings.Memorization },
-        { label: strings.Revision, value: strings.Revision },
-        { label: strings.Reading, value: strings.Reading },
-      ];
-      let selectedAssignmentTypeIndex = 0;
-      if (this.props.assignmentType !== undefined) {
-        if (
-          options.findIndex(
-            option => option.value === this.props.assignmentType
-          ) !== -1
-        ) {
-          selectedAssignmentTypeIndex = options.findIndex(
-            option => option.value === this.props.assignmentType
-          );
-        }
-      }
-
       return (
         <View style={{ width: screenWidth, height: screenHeight }}>
           <ScrollView
@@ -478,27 +318,6 @@ export default class MushafScreen extends QcParentScreen {
               {this.state.pages.map((item, idx) => this.renderItem(item, idx))}
             </Swiper>
           </ScrollView>
-          <View style={{ padding: 5 }}>
-            {this.state.selection.start.surah > 0 ||
-            this.state.freeFormAssignment ? (
-              <Text style={fontStyles.mainTextStyleDarkGrey}>
-                {this.state.assignmentName}
-              </Text>
-            ) : (
-              <View />
-            )}
-          </View>
-          <SwitchSelector
-            options={options}
-            initial={selectedAssignmentTypeIndex}
-            height={20}
-            textColor={colors.darkGrey}
-            selectedColor={colors.primaryDark}
-            buttonColor={colors.primaryLight}
-            borderColor={colors.lightGrey}
-            onPress={value => this.setState({ assignmentType: value })}
-            style={{ marginTop: 2 }}
-          />
         </View>
       );
     }
